% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_tensor}
\alias{as_tensor}
\title{Ensure Input is a Torch Tensor}
\usage{
as_tensor(x, dtype = torch_float())
}
\arguments{
\item{x}{A matrix, data frame, vector, or torch tensor.}

\item{dtype}{Torch data type (default: torch_float()).}
}
\value{
A torch tensor.
}
\description{
Ensure Input is a Torch Tensor
}
\keyword{internal}
