% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dftomat}
\alias{dftomat}
\title{Convert Data Frame to Numeric Matrix}
\usage{
dftomat(X)
}
\arguments{
\item{X}{A data frame to be converted.}
}
\value{
A numeric matrix.
}
\description{
Converts a data frame into a numeric matrix. Factor and character
variables are converted to numeric.
}
\keyword{internal}
