% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frengression_seq.R, R/frengression_surv.R,
%   R/predict.R
\name{predict_causal.frengression_seq}
\alias{predict_causal.frengression_seq}
\alias{predict_causal.frengression_surv}
\alias{predict_causal}
\alias{predict_causal.frengression}
\title{Predict Causal Effects (generic)}
\usage{
\method{predict_causal}{frengression_seq}(model, s, x, target = "mean", sample_size = 100, ...)

\method{predict_causal}{frengression_surv}(model, s, x, sample_size = 100, ...)

predict_causal(model, ...)

\method{predict_causal}{frengression}(model, x, target = "mean", sample_size = 100, ...)
}
\arguments{
\item{model}{A frengression object.}

\item{s}{Static baseline variables (matrix or tensor).}

\item{x}{Intervention values.}

\item{target}{"mean" or "sample" (default: "mean").}

\item{sample_size}{Number of Monte Carlo samples (default: 100).}

\item{...}{Additional arguments passed to methods.}
}
\value{
A matrix of predictions.
}
\description{
Predict Causal Effects (generic)
}
