% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.frengression}
\alias{print.frengression}
\title{Print a Frengression Model}
\usage{
\method{print}{frengression}(x, ...)
}
\arguments{
\item{x}{A frengression object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print a Frengression Model
}
