% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.frengression_seq}
\alias{print.frengression_seq}
\title{Print a FrengressionSeq Model}
\usage{
\method{print}{frengression_seq}(x, ...)
}
\arguments{
\item{x}{A frengression_seq object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print a FrengressionSeq Model
}
