% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.frengression_surv}
\alias{print.frengression_surv}
\title{Print a FrengressionSurv Model}
\usage{
\method{print}{frengression_surv}(x, ...)
}
\arguments{
\item{x}{A frengression_surv object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print a FrengressionSurv Model
}
