% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfrengression-package.R
\docType{package}
\name{Rfrengression-package}
\alias{Rfrengression}
\alias{Rfrengression-package}
\title{Rfrengression: Frugal, Flexible, Faithful Causal Data Simulation via Frengression}
\description{
Deep generative models for causal data simulation. Learns joint distributions of covariates, treatments, and outcomes, enabling interventional sampling via do-calculus. Implements three model classes: cross-sectional (Frengression), longitudinal (FrengressionSeq), and survival (FrengressionSurv). Based on "Frugal, Flexible, Faithful: Causal Data Simulation via Frengression" by Shen and Meinshausen (2025) \href{https://arxiv.org/abs/2508.01018}{arXiv:2508.01018}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://xiangao.github.io/Rfrengression/}
  \item \url{https://github.com/xwshen51/frengression/}
  \item Report bugs at \url{https://github.com/xwshen51/frengression/issues}
}

}
\author{
\strong{Maintainer}: Nicolai Meinshausen \email{meinshausen@stat.math.ethz.ch}

Authors:
\itemize{
  \item Nicolai Meinshausen \email{meinshausen@stat.math.ethz.ch}
  \item Xinwei Shen \email{xinwei.shen@stat.math.ethz.ch}
}

}
\keyword{internal}
